requirements.txt
# AI Demystified Starter Kit — Dependencies
# Install with: pip install -r requirements.txt

# Web UI
streamlit>=1.32.0

# LLM Providers
openai>=1.12.0
groq>=0.4.2

# Vector DB (for future RAG)
pinecone>=3.0.0

# Environment variable management
python-dotenv>=1.0.0

# Network libraries (pin httpx to avoid "proxies" kwarg error in openai/groq)
httpx<0.28.0
